+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds the `"editorMode"` option (pencil) to the `Icon` component.
+- Adds the `Errors` page to the `Accessibility Guide` section of Storybook.
 
 ## 3.0.0 (March 14, 2024) React 18 / Chakra 2.8