+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds the `"editorMode"` option (pencil) to the `Icon` component.
 
+### Updates
+
+- Updates default heading size to `heading4` on Modal component.
+
 ## 3.0.0 (March 14, 2024) React 18 / Chakra 2.8
 
 ### Adds